almost done

+++ springboot-e-commerce/src/main/java/com/ecommerce/service/ProductService.java
@@ -25,6 +25,10 @@ public class ProductService {
 		return this.productRepository.findAll();
 	}
 	
+	public Optional<Product> getProductById(Long id) {
+		return this.productRepository.findById(id);
+	}
+	
 	public Product save(Product product) {
 		Optional<User> user = userRepository.findById(product.getUserId());
 		if(user.isPresent()) {
@@ -36,20 +40,20 @@ public class ProductService {
 	
 	public Product edit(long id, Product editProduct) {
 		Optional<Product> productData = productRepository.findById(id);
-		Product product = productData.get();
-		//4
-		
-//		product.set(editProduct);
-		product.setName(editProduct.getName());
-		product.setDecription(editProduct.getDecription());
-		product.setQuantity(editProduct.getQuantity());
-		product.setPrice(editProduct.getPrice());
-		return productRepository.save(product);
+		if(productData.isPresent()) {
+			Product product = productData.get();
+			product.set(editProduct);		
+			return productRepository.save(product);
+		}
+		return null;
 	}
 	
 	public Boolean deleteById(Long id) {
-		//exist
-        productRepository.deleteById(id);
-        return true;
+		Optional<Product> product = productRepository.findById(id);
+		if(product.isPresent()) {
+	        productRepository.deleteById(id);
+	        return true;
+		}
+		return false;
 	}
 }
